added part

+++ src/commands/nick.cpp
@@ -13,7 +13,7 @@ void Server::parseNick(std::string buffer, Client &client) {
         print_err(this->getHostname() + " 432 " + buffer + " :Erroneus nickname");
         return ;
     }
-    for (int i = 0; i < this->getMap().size(); i++)
+    for (size_t i = 0; i < this->getMap().size(); i++)
     {
         if (this->map[i].getNick().compare(buffer) == 0){
             print_err(this->getHostname() + " 433 " + buffer + " :Nickname is already in use");
